chore: rename backend to platfrom
