+++ apps/platform/middleware/03.auth.ts
