+++ apps/platform/nitro.config.ts
