+++ apps/platform/routes/auth/logout.post.ts
