+++ apps/platform/routes/auth/status.get.ts
