+++ apps/platform/routes/health.get.ts
