+++ apps/platform/routes/trpc/[trpc].ts
