+++ apps/platform/utils/auth.ts
