+++ apps/platform/utils/signup.ts
