+++ apps/platform/utils/verifyDns.ts
